more hai functions
