+++ wafel_core/include/wafel/hai.h
@@ -7,4 +7,8 @@ LINKABLE void hai_redirect_mlc2sd(void);
 
 LINKABLE void hai_companion_add_offset(uint32_t *buffer, uint32_t offset);
 
+LINKABLE void hai_apply_getdev_patch(void);
+
+LINKABLE int hai_getdev(void);
+
 #endif
